23-03

+++ app/Http/Controllers/PostController.php
@@ -23,45 +23,36 @@ function almacenar_post($request, $estado){
 
 class PostController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index(Request $request)
-    {
-        $posts = Post::paginate(5);
-        if($request -> ajax()){
-            $view = view('data',compact('posts'))->render();
-            return response()->json(['html' => $view]);
-        }
-        return view('post',compact('posts'));
-    }
+    
+    // public function index(Request $request)
+    // {
+    //     $posts = Post::paginate(5);
+    //     if($request -> ajax()){
+    //         $view = view('data',compact('posts'))->render();
+    //         return response()->json(['html' => $view]);
+    //     }
+    //     return view('post',compact('posts'));
+    // }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
+    // public function create()
+    // {   
+    //     //crea un nuevo post en blanco
+    //     $post_crear = Post::create([
+    //         'id_user'=>auth()->user()->id,
+    //         'tipo'=>'publicacion',
+    //         'estado' =>'editando'
+    //     ]);
+        
+    //     $aulas = Aula::select('*')->where('aula',"!=", 'No asignado')->get();
+    //     return view('admin.post.create')->with('post',$post_crear)->with('aulas',$aulas); 
+    // }
+
+    public function create($tipo)
     {   
-        //Elimina los posts que no fueron ni publicados ni borrador
-        // $posts = Post::select('*')->where('estado', 'eliminar')->where('id_user', auth()->user()->id)->get();
-        // foreach($posts as $post){
-        //     foreach($post->archivos as $archivo){
-        //         $archivo->delete();
-        //     }
-        //     if($post->carpeta != null){
-        //         Storage::disk("google")->deleteDirectory($post->carpeta);
-        //     }
-            
-        //     $post->delete();
-        // }
-
         //crea un nuevo post en blanco
         $post_crear = Post::create([
             'id_user'=>auth()->user()->id,
-            'tarea'=>'1',
+            'tipo'=>$tipo,
             'estado' =>'eliminar'
         ]);
         
@@ -69,13 +60,17 @@ class PostController extends Controller
         return view('admin.post.create')->with('post',$post_crear)->with('aulas',$aulas); 
     }
 
-    public function create_profesor($id)
+    public function createPostCurso($id,$tipo)
     {   
-        $curso = Curso::find($id);
-        $profesor = Profesor::find($curso->id_profesor);
-        $aulas = Aula::select('*')->where('aula', 'No asignado')->get();
-        return view('admin.post.create')->with('curso',$curso)->with('profesor',$profesor)->with('aulas',$aulas);
- 
+        $curso_post = Curso::find($id);
+        $alumnos = Alumno::select('*')->where('id_aula',$curso_post->aula->id)->get();
+        $post_crear = Post::create([
+            'id_user'=>auth()->user()->id,
+            'tipo'=>$tipo,
+            'estado' =>'eliminar'
+        ]);
+        $aulas = Aula::select('*')->where('aula',"!=", 'No asignado')->get();
+        return view('admin.post.create')->with('post',$post_crear)->with('curso_post',$curso_post)->with('alumnos',$alumnos)->with('aulas',$aulas);
     }
 
     public function store(Request $request)
@@ -99,11 +94,11 @@ class PostController extends Controller
         }
         //establece la hora de publicacion segun el estado del post
         if($request->get('estado')=='publicar'){
-            $datos['fecha_hora'] = date("Y-m-d H:i:s");
+            $datos['fecha_publicacion'] = date("Y-m-d H:i:s");
         }elseif($request->get('estado')=='borrador'){
-            $datos['fecha_hora'] = null;
+            $datos['fecha_publicacion'] = null;
         }elseif($request->get('estado')=='programar'){
-            $datos['fecha_hora'] = $request->get('fecha')." ". $request->get('hora');
+            $datos['fecha_publicacion'] = $request->get('fecha')." ". $request->get('hora');
         }   
         
         //actualiza el post
@@ -139,7 +134,7 @@ class PostController extends Controller
 
         if(auth()->user()->role=='admin'){
             if($request->id_curso){
-                return redirect()->route('curso.perfil',$request->id_curso);   
+                return redirect()->route('admin.curso.perfil',$request->id_curso);   
             }else{
                 return redirect()->route('admin.inicio');
             }
@@ -150,7 +145,7 @@ class PostController extends Controller
         }
     }
 
-    public function edit($id)
+    public function edit($tipo,$id)
     {
         $aulas_checked = [];
         
@@ -186,13 +181,14 @@ class PostController extends Controller
         
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request)
     {      
         $request->validate([
             'imagen' => 'image|mimes:jpeg,png,jpg'
         ]);
+
         $datos = $request->all();
-        $post = Post::find($id);
+        $post = Post::find($request->get('id_post'));
         if($imagen = $request->file('imagen')) {         
             $nombre_img = date('YmdHis'). "." . $request->file('imagen')->getClientOriginalExtension();
             $imagen = $request->file('imagen')->storeAs('imagenes_post',$nombre_img,'public');
@@ -203,16 +199,24 @@ class PostController extends Controller
 
         //establece la hora de publicacion segun el estado del post
         if($request->get('estado')=='publicar' && $post->estado =='publicar'){
-            $datos['fecha_hora'] = $post->fecha_hora;
+            $datos['fecha_publicacion'] = $post->fecha_publicacion;
         }elseif($request->get('estado')=='borrador'){
-            $datos['fecha_hora'] = null;
+            $datos['fecha_publicacion'] = null;
         }elseif($request->get('estado')=='programar'){
-            $datos['fecha_hora'] = $request->get('fecha')." ". $request->get('hora');
+            $datos['fecha_publicacion'] = $request->get('fecha')." ". $request->get('hora');
         }       
+        if($request->get('recibir')==null){
+            $datos['recibir'] = 'no';
+        }
+        if($request->get('id_curso')==null){
+            $datos['id_curso'] = null;
+        }
+
+
         $post->update($datos);
 
 
-        $post_alumnos = AlumnoPost::select('*')->where('id_post', $id)->get();
+        $post_alumnos = AlumnoPost::select('*')->where('id_post', $request->get('id_post'))->get();
         foreach($post_alumnos as $post_alumno){
             $post_alumno->delete();
         }
@@ -221,21 +225,21 @@ class PostController extends Controller
         if(count($request->aulas)>1){
             foreach($request->aulas as $aula){
                 AlumnoPost::create([
-                    'id_post'=>$id,
+                    'id_post'=>$request->get('id_post'),
                     'id_aula'=>$aula
                 ]);
             }
         }else{
             if(in_array('0',$request->alumnos)){
                 AlumnoPost::create([
-                    'id_post'=>$id,
+                    'id_post'=>$request->get('id_post'),
                     'id_aula'=>($request->aulas)[0],               
                 ]);
             }
             else{
                 foreach($request->alumnos as $alumno){
                     AlumnoPost::create([
-                        'id_post'=>$id,
+                        'id_post'=>$request->get('id_post'),
                         'id_aula'=>($request->aulas)[0],
                         'id_curso'=>$request->get('id_curso'),
                         'id_alumno'=>$alumno
@@ -256,13 +260,13 @@ class PostController extends Controller
 
         if(auth()->user()->role=='admin'){
             if($post->id_curso){
-                return redirect()->route('curso.perfil',$post->id_curso);   
+                return redirect()->route('admin.curso.perfil',$post->id_curso);   
             }else{
                 return redirect()->route('admin.inicio');
             }           
         }
         else if(auth()->user()->role=='profesor'){
-            return redirect()->route('curso.perfil',$post->id_curso);
+            return redirect()->route('profesor.curso.perfil',$post->id_curso);
         }
     }
 
@@ -270,21 +274,24 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        Storage::delete(str_replace("storage", "public", $post->imagen)); 
-        foreach($post->archivos as $archivo){
-            $archivo->delete();
-        }
 
+        if($post->imagen != null){
+            Storage::delete(str_replace("storage", "public", $post->imagen)); 
+        }
+        
         $post_alumnos = AlumnoPost::select('*')->where('id_post',$id)->get();
         foreach($post_alumnos as $post_alumno){
             $post_alumno->delete();
         }
 
-        Storage::disk("google")->deleteDirectory($post->carpeta);
-
+        if($post->carpeta != null){
+            Storage::disk("google")->deleteDirectory($post->carpeta);
+            foreach($post->archivos as $archivo){
+                $archivo->delete();
+            }
+        }
 
         $post->delete();  
-
 
         return redirect()->back();
     }
@@ -320,8 +327,8 @@ class PostController extends Controller
     public function eliminar_post_crear(Request $request)
     {   
         //Elimina los posts que no fueron ni publicados ni borrador
-        $posts = Post::select('*')->where('estado', 'eliminar')->where('id_user',$request->get('id_user'))->get();
-        foreach($posts as $post){
+        $post = Post::select('*')->where('estado', 'eliminar')->where('id',$request->get('id_post'))->first();
+        
             foreach($post->archivos as $archivo){
                 $archivo->delete();
             }
@@ -330,7 +337,7 @@ class PostController extends Controller
             }
             
             $post->delete();
-        }
+        
         return response()->json([      
             'html1'=>'hola',
         ]);
